% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNStatistic.R
\name{buildProbabilityTreeOnTargetGene}
\alias{buildProbabilityTreeOnTargetGene}
\title{Find all genes that are related to the target gene}
\usage{
buildProbabilityTreeOnTargetGene(targetGene, currentState, previousState,
  timeseries, genes, matchedgenes, matchedexpression, maxK = 5,
  temporal = 1, allRelatedTargetGeneMeasures,
  allRelatedConditionalGenesMeasures, useParallel = FALSE)
}
\arguments{
\item{targetGene}{The target gene}

\item{currentState}{current combined gene state i.e. the gene state at t}

\item{previousState}{previous combined gene state i.e. the gene state at t-1}

\item{genes}{a list of genes that the related genes re mined from}

\item{matchedgenes}{a list of pre matched genes}

\item{matchedexpression}{previous matched expression string}

\item{maxK}{The maximum depth the cube can go to. The default value is 5}

\item{temporal}{1: d means temporal = t-1 ... t-d}

\item{allRelatedTargetGeneMeasures}{all measures that are related to the target genes.}

\item{allRelatedConditionalGenesMeasures}{all measures that are related to the conditional genes}

\item{timeseriesCube}{a list of samples in which a sample contains gene and time points}

\item{An}{option to turn on parallel}
}
\value{
A list of measures
}
\description{
Find all genes that are related to the target gene
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
completeState<-extractGeneStateFromTimeSeriesCube(listtest)
getCurrentState<-completeState
getpreviousState<-completeState
getCurrentState<-getCurrentState[,-1]#remove the first one
getpreviousState<-getpreviousState[,-length(colnames(getpreviousState))] #remove the last one
clusteringOnTargetGene("gene1",getCurrentState,getpreviousState,listtest,c("gene1","gene2","gene3"),1,FALSE)
}
