% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNStatistic.R
\name{clusteringOnTargetGene}
\alias{clusteringOnTargetGene}
\title{Find all genes that are related to the target gene}
\usage{
clusteringOnTargetGene(targetGene, currentState, previousState, timeseriesCube,
  genes, temporal = 1, useParallel = FALSE)
}
\arguments{
\item{targetGene}{The target gene}

\item{currentState}{current combined gene state i.e. the gene state at t}

\item{previousState}{previous combined gene state i.e. the gene state at t-1}

\item{timeseriesCube}{a list of samples in which a sample contains gene and time points}

\item{genes}{a list of genes that the related genes re mined from}

\item{temporal}{1: d means temporal = t-1 ... t-d}
}
\value{
A list of genes that are related to the target genes
}
\description{
Find all genes that are related to the target gene
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
completeState<-extractGeneStateFromTimeSeriesCube(listtest)
getCurrentState<-completeState
getpreviousState<-completeState
getCurrentState<-getCurrentState[,-1]#remove the first one
getpreviousState<-getpreviousState[,-length(colnames(getpreviousState))] #remove the last one
clusteringOnTargetGene("gene1",getCurrentState,getpreviousState,listtest,c("gene1","gene2","gene3"),1,FALSE)
}
