% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNCube.R
\name{constructFBNCube}
\alias{constructFBNCube}
\title{Create an Orchard cube}
\usage{
constructFBNCube(targetgenes, allgenes, timeseriesCube, maxK = 5,
  temporal = 1, useParallel = FALSE)
}
\arguments{
\item{targetgenes}{A vector of genes that will be treated as target genes}

\item{allgenes}{All genes that are available for building up the cube}

\item{timeseriesCube}{A list of samples in which a sample is a matrix that contains gene states where genes in rows and time points in columms}

\item{maxK}{The maximum level the cube can dig in}

\item{temporal}{A value that used to be 1 indicates the previous steps the current one can depend on}

\item{useParallel}{If it is TRUE, the constructing will run it in parallel, otherwise in a singl thread}
}
\value{
A Orcahrd cube that contains all precomputed measures
}
\description{
Create an Orchard cube
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
constructFBNCube(c("gene1","gene2"),c("gene1","gene2","gene3"),listtest,4,1,FALSE)
}
