% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNCore.R
\name{extractGeneStateFromTimeSeriesCube}
\alias{extractGeneStateFromTimeSeriesCube}
\title{An internal method to extract gene state from timeseries cube}
\usage{
extractGeneStateFromTimeSeriesCube(timeSeriesCube)
}
\arguments{
\item{timeSeriesCube}{A list object contains samples in which a sample contains genes in row and time points in column}
}
\value{
A combined matrix
}
\description{
An internal method to extract gene state from timeseries cube
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
extractGeneStateFromTimeSeriesCube(listtest)
}
