% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNModelProcessor.R
\name{FBNBenchmark}
\alias{FBNBenchmark}
\title{A benchmark type function to test a complete process of FBN model}
\usage{
FBNBenchmark(booleanNetwork, shorttrainingtimeSeries, shorttestseries,
  originaltestseries, interval = 1, temporal = 1, convertToFBN = TRUE,
  existingFBNCube = NULL)
}
\arguments{
\item{booleanNetwork}{A Boolean Network object}

\item{shorttrainingtimeSeries}{A training timeseries matrix data set}

\item{shorttestseries}{A test timeseries matrix data set}

\item{originaltestseries}{A original timeseries matrix data set}

\item{interval}{The internal gat between time points}

\item{temporal}{A temporal setting with default =1}

\item{convertToFBN}{if it is a need to convert the boolean network to FBN Network. If the boolean network object is not a FBNNetwork object, then it must be set to TRUE}

\item{existingFBNCube}{An existing FBN cube object}

\item{decay}{The decay time step for protein decay}
}
\value{
A list of object
}
\description{
A benchmark type function to test a complete process of FBN model
}
\examples{
coming later
}
