\name{FBNNet-package}
\alias{FBNNet-package}
\alias{FBNNet}
\docType{package}
\title{
\packageTitle{FBNNet}
}
\description{
\packageDescription{FBNNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FBNNet}
\packageIndices{FBNNet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FBNNet}

Maintainer: \packageMaintainer{FBNNet}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
