% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNGraph.R
\name{FBNNetwork.Graph.ConvertToNetworkGraphicObject}
\alias{FBNNetwork.Graph.ConvertToNetworkGraphicObject}
\title{This method is used to convert the FBN network into a FBN networkobject for graphs}
\usage{
FBNNetwork.Graph.ConvertToNetworkGraphicObject(FBNnetwork)
}
\arguments{
\item{FBNnetwork}{A FBN network object}
}
\value{
An network graphic object
}
\description{
This method is used to convert the FBN network into a FBN networkobject for graphs
}
\examples{
coming later
}
