% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNGraph.R
\name{FBNNetwork.Graph.StaticNetwork}
\alias{FBNNetwork.Graph.StaticNetwork}
\title{Display a static network in a slice}
\usage{
FBNNetwork.Graph.StaticNetwork(FBNnetwork)
}
\arguments{
\item{FBNnetwork}{A FBN network object}
}
\value{
No return
}
\description{
Display a static network in a slice
}
\examples{
coming later
}
