% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNCore.R
\name{getGenePrababilities}
\alias{getGenePrababilities}
\title{A core method to calculate all measures for a target gene with a condition gene}
\usage{
getGenePrababilities(timeSeriesCube, fixedgenestate, targetgene, newgene,
  geneStates = NULL, preGeneStates = NULL, temporal = 1,
  temporalGeneStates = NULL)
}
\arguments{
\item{timeSeriesCube}{A list object contains samples in which a sample contains genes in row and time points in column}

\item{fixedgenestate}{prefixed gene state}

\item{targetgene}{The tage gene's name}

\item{newgene}{The new added conditional gene}

\item{geneStates}{The current combined gene state}

\item{preGeneStates}{The previous combined gene state}

\item{temporal}{The temporal setting and the default value is 1}

\item{temporalGeneStates}{The combined gene state for temporal}
}
\value{
A list of measures
}
\description{
A core method to calculate all measures for a target gene with a condition gene
}
\examples{
genesInput<-c("CycD","p27","CycE","E2F")
testseries<-list()
testseries[[1]]<-matrix(c(1,1,1,1,0,1,0,1,1,1,1,1,0,0,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))
testseries[[2]]<-matrix(c(1,1,0,1,0,1,0,1,0,1,0,1,1,0,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))
testseries[[3]]<-matrix(c(1,0,0,1,1,1,0,1,1,1,0,1,1,1,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))

getCurrentState<-extractGeneStateFromTimeSeriesCube(testseries)
getCurrentState<-getCurrentState[,-1]#remove the first one

getpreviousState<-extractGeneStateFromTimeSeriesCube(testseries)
getpreviousState<-getpreviousState[,-length(colnames(getpreviousState))] #remove the last one
probability<-getGenePrababilities(testseries,NULL,"CycD","p27",getCurrentState,getpreviousState,1)
}
