% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNDataUtility.R
\name{getRelatedAffyRawData}
\alias{getRelatedAffyRawData}
\title{A benchmark type function to test a complete process of FBN model
Step 1, read Affy data}
\usage{
getRelatedAffyRawData(cellDirectory)
}
\description{
A benchmark type function to test a complete process of FBN model
Step 1, read Affy data
}
