% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNDataUtility.R
\name{identifyDifferentiallyExpressedGenes}
\alias{identifyDifferentiallyExpressedGenes}
\title{A benchmark type function to test a complete process of FBN model
Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes}
\usage{
identifyDifferentiallyExpressedGenes(orderSampleTimeSeries)
}
\description{
A benchmark type function to test a complete process of FBN model
Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes
}
