% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNMiner.R
\name{mineFBNNetwork}
\alias{mineFBNNetwork}
\title{Mine FBN Networks from an Orchard cube}
\usage{
mineFBNNetwork(fbnGeneCube, genes, useParallel = FALSE)
}
\arguments{
\item{fbnGeneCube}{A pre constructed Orchard cube}

\item{genes}{The target genes in the output}

\item{useParallel}{An option turns on parallel}
}
\value{
A object of FBN network
}
\description{
Mine FBN Networks from an Orchard cube
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
cube<-constructFBNCube(c("gene1","gene2"),c("gene1","gene2","gene3"),listtest,4,1,FALSE)
network<-mineFBNNetwork(cube,c("gene1","gene2"))
}
