% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNDataUtility.R
\name{normalizeTimesereisRawData}
\alias{normalizeTimesereisRawData}
\title{A benchmark type function to test a complete process of FBN model
Step 2, normalizing data}
\usage{
normalizeTimesereisRawData(rawData)
}
\description{
A benchmark type function to test a complete process of FBN model
Step 2, normalizing data
}
