% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNModel.R
\name{reconstructTimeseries}
\alias{reconstructTimeseries}
\title{This method is used to reconstruct time series data}
\usage{
reconstructTimeseries(fbnCube, fbnnetwork, timeseriesCube, interval = 1,
  maxTimepoints = 100)
}
\arguments{
\item{fbnCube}{A pre constructed Orchard cube}

\item{fbnnetwork}{An object of FBNNetwork}

\item{timeseriesCube}{An original time series data that contains samples, genes and time points. It can be as list of samples contain genes and the initial time point.}

\item{interval}{An interval is the time gap between two time points}
}
\value{
A list object that contains reconstructed time series and FBN network
}
\description{
This method is used to reconstruct time series data
}
\examples{
mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtestOriginal<-list(mat1,mat2)
cube<-constructFBNCube(c("gene1","gene2"),c("gene1","gene2","gene3"),listtestOriginal,4,1,FALSE)
network<-mineFBNNetwork(cube,c("gene1","gene2"))
mat1<-matrix(c("1","1","0"),3,1, dimnames=list(c("gene1","gene2","gene3"),c("1")))
mat2<-matrix(c("1","1","1"),3,1, dimnames=list(c("gene1","gene2","gene3"),c("1")))
listtestInitial<-list(mat1,mat2)
result<-reconstructTimeseries(cube,network,listtestInitial,1,1,3)
}
