% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNUtility.R
\name{splitExpression}
\alias{splitExpression}
\title{Split a compressed gene regulate function into symbols}
\usage{
splitExpression(expression, outputType = 1, lowerCase = FALSE)
}
\arguments{
\item{expression}{A regulate function like GeneA&GeneB&!GeneC}

\item{outputType}{A output type with two options 1 the splition included "!", others not included "!"}

\item{lowerCase}{Is lower case required?}
}
\value{
The return is a vector of the input regulate function, a example is "GeneA" "&" "GeneB" "&" "!" "GeneC"
}
\description{
Split a compressed gene regulate function into symbols
}
\examples{
splitExpression("GeneA&GeneB&!GeneC")
splitExpression("GeneA&GeneB&!GeneC",1)
splitExpression("GeneA&GeneB&!GeneC",0)
}
