% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNUtility.R
\name{timeseriesReduction}
\alias{timeseriesReduction}
\title{Remove some time points out from the target time series}
\usage{
timeseriesReduction(timeseriesCube, reducedTimePoints)
}
\arguments{
\item{timeseriesCube}{A data set that contains samples and each sample contains genes and time points}

\item{reducedTimePoints}{How many time points reduced between two observation time points}
}
\value{
reduced time series cube
}
\description{
Remove some time points out from the target time series
}
\examples{
coming later
}
